+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Handle Twitter names that are too long properly
+
 ## [0.10.3] - 2022-03-13
 
 ### Fixed